Give analyzers the ability to report failures without raising exceptions that stop the run entirely
This is useful for the Go analyzer since it can fail to analyze some of the
source code using backvendor, but it should continue the analysis. So now,
instead of just failing outright, this PR makes it so that the analyzers
can choose to continue but denote to the caller that it wasn't a total success
with an AnalysisFailure exception.

Addresses FACTORY-3520